move notify and interactions into new scheduler module
